[REFACTOR] 변경된 경로의 서비스를 import하는 클래스의 경로 수정
