+++ src/main/java/com/kakaobase/snsapp/domain/posts/scheduler/PostCacheSyncScheduler.java
@@ -1,7 +1,7 @@
 package com.kakaobase.snsapp.domain.posts.scheduler;
 
 
-import com.kakaobase.snsapp.domain.posts.service.PostCacheSyncService;
+import com.kakaobase.snsapp.domain.posts.service.cache.PostCacheSyncService;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.scheduling.annotation.Scheduled;
